Python/requirements.txt
mysql-connector
schedule
telepot
python-telegram-bot
utils
